# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ExerciseDB API
  version: 0.1.0
  description: >
    A public exercise catalog API for fitness app developers.


    Every endpoint returns a `{ success, data }` envelope. Errors are
    [RFC 9457 Problem Details](https://www.rfc-editor.org/rfc/rfc9457) served as
    `application/problem+json` — branch on the `code` member, which is stable.


    Authenticate with an API key in the `x-api-key` header, or as
    `Authorization: Bearer <key>`. Keys are issued once at registration and are
    never shown again.


    Quotas are per API key and reset at midnight UTC. Every response carries
    `X-RateLimit-Limit`, `X-RateLimit-Remaining`, and `X-RateLimit-Reset`.


    Every response also carries `X-Request-Id`, repeated as `requestId` in any
    error body. Send your own `X-Request-Id` to trace a call end to end.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
...
  /auth/register:
    post:
      summary: Register a developer account and issue the first API key
      description: >
        The plaintext key in `data.apiKey.key` is returned exactly once. The
        server stores only a SHA-256 hash and cannot show it to you again.
        A browser session cookie is set at the same time.
      operationId: registerDeveloper
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created, with a one-time plaintext API key.
          headers:
            Set-Cookie:
              description: '`exdb_session=…; HttpOnly; SameSite=Lax; Path=/`'
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/login:
    post:
      summary: Log in and start a browser session
      description: >
        Sets an httpOnly `exdb_session` cookie. It issues **no** API key —
        earlier versions minted one on every login, which accumulated live
        credentials. Create keys explicitly with `POST /me/keys`.
      operationId: loginDeveloper
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated. A session cookie is set.
          headers:
            Set-Cookie:
              description: '`exdb_session=…; HttpOnly; SameSite=Lax; Path=/`'
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      summary: End the current browser session
      description: >
        Revokes the session server-side and clears the cookie. A token captured
        before logout stops working immediately. Safe to call when signed out.
      operationId: logoutDeveloper
      tags: [Auth]
      security: []
      responses:
        '200':
          description: The session is revoked and the cookie cleared.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /me:
    get:
      summary: Get the current developer account
      operationId: getCurrentDeveloper
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - SessionCookie: []
      tags: [Auth]
      responses:
        '200':
          description: The account the API key belongs to.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeveloperAccountResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /me/keys:
    get:
      summary: List API keys for the current developer account
      description: Metadata only. Plaintext key values are never returned here.
      operationId: listDeveloperApiKeys
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - SessionCookie: []
      tags: [Auth]
      responses:
        '200':
          description: API key metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Create a new API key
      operationId: createDeveloperApiKey
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - SessionCookie: []
      tags: [Auth]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
      responses:
        '201':
          description: The new key, including its one-time plaintext value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuedApiKeyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /me/keys/{id}:
    delete:
      summary: Revoke an API key
      operationId: revokeDeveloperApiKey
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - SessionCookie: []
      tags: [Auth]
      parameters:
        - $ref: '#/components/parameters/ApiKeyId'
      responses:
        '200':
          description: The revoked key's metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /me/usage:
    get:
      summary: Get daily usage for the current developer account
      operationId: getDeveloperUsage
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - SessionCookie: []
      tags: [Auth]
      responses:
        '200':
          description: One row per day, most recent first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /sync/metadata:
    get:
      summary: Get catalog version and sync limits
      description: >
        Call this first. `catalogVersion` and `latestChangeAt` tell you whether
        a sync is worth running at all.
      operationId: getSyncMetadata
      tags: [Sync]
      responses:
        '200':
          description: Catalog version, latest timestamps, resources, and enums.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncMetadataResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /sync/exercises:
    get:
      summary: Sync exercises changed since a timestamp
      description: >
        Omit `updated_since` for a first full sync. Follow
        `pagination.nextCursor` until `hasMore` is false, then commit
        `data.latestChangeAt` and send it as `updated_since` next time.
        `limit` bounds change events, not exercises, so a full page may return
        fewer than `limit` records — never page on `exercises.length`.
      operationId: syncExercises
      tags: [Sync]
      parameters:
        - $ref: '#/components/parameters/UpdatedSince'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
        - name: cursor
          in: query
          description: Opaque cursor from the previous page. Never construct one.
          schema: { type: string }
        - $ref: '#/components/parameters/IncludeDeprecated'
      responses:
        '200':
          description: Changed exercises, tombstones, and a cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncExercisesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Your API key, prefixed `exdb_`.
    BearerAuth:
      type: http
      scheme: bearer
      description: 'The same API key, sent as `Authorization: Bearer exdb_…`.'
    SessionCookie:
      type: apiKey
      in: cookie
      name: exdb_session
      description: >
        Browser session set by `/auth/login` and `/auth/register`. httpOnly,
        so JavaScript cannot read it. Used by the developer dashboard on
        `/me/*` and `/billing/checkout`; it consumes no daily quota. Machine
        callers should use an API key instead.
  parameters:
...
    UpdatedSince:
      name: updated_since
      in: query
      description: ISO 8601 timestamp. Only records changed after it are returned.
      schema:
        type: string
        format: date-time
        example: '2026-06-15T10:00:00.000Z'
    IncludeDeprecated:
      name: include_deprecated
      in: query
      description: >
        Include deprecated records, for migration and cleanup flows. On
        `/sync/exercises` this also suppresses their `deprecated` tombstones:
        the record is returned with `status: deprecated` instead.
      schema: { type: boolean, default: false }
...
    ApiKeyId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: The API key is missing, malformed, revoked, or expired.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: >
        Authenticated, but not permitted. Premium exercises require a `pro` or
        `enterprise` tier; inactive accounts cannot log in.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: No such resource.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: The request conflicts with existing state.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    TooManyRequests:
      description: The daily quota for this API key is exhausted.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema: { type: integer }
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    InternalServerError:
      description: >
        An unexpected error. `detail` is always a generic message; internals are
        never exposed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          enum: [true]
...
    CursorPagination:
      type: object
      required: [limit, hasMore]
      properties:
        limit: { type: integer, example: 100 }
        nextCursor:
          type: string
          nullable: true
          description: Pass back as `cursor`. Null when there are no more pages.
          example: eyJvZmZzZXQiOjEwMH0
        hasMore: { type: boolean, example: true }

    # ---------------------------------------------------------------------
    # Core catalog models
    # ---------------------------------------------------------------------
    ExerciseSummary:
      type: object
      description: The fields needed to render an exercise in a list.
      required:
        [id, slug, name, status, difficulty, movementPattern, tags, updatedAt]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string, example: barbell-back-squat }
        name: { type: string, example: Barbell Back Squat }
        status:
          type: string
          enum: [draft, active, deprecated]
        category:
          type: string
          nullable: true
          description: Category slug.
          example: strength
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        movementPattern:
          type: string
          enum: [squat, hinge, push, pull, carry, rotation, gait]
        tags:
          type: array
          items: { type: string }
          example: [compound, lower-body]
        updatedAt: { type: string, format: date-time }
    Exercise:
      description: A full exercise record, including coaching and programming data.
      allOf:
        - $ref: '#/components/schemas/ExerciseSummary'
        - type: object
          required:
            [
              description,
              instructions,
              tips,
              contraindications,
              mechanics,
              jointRegions,
              flags,
              programming,
              isPremium,
              catalogVersion
            ]
          properties:
            description: { type: string }
            instructions:
              type: array
              items: { type: string }
              description: Ordered steps.
            tips:
              type: array
              items: { type: string }
            breathingCues:
              type: string
              nullable: true
            contraindications:
              type: array
              items: { type: string }
            forceType:
              type: string
              nullable: true
              enum: [push, pull, static, compound]
            mechanics:
              type: string
              enum: [compound, isolation]
            position:
              type: string
              nullable: true
              enum: [standing, seated, lying, kneeling, other]
            planeOfMotion:
              type: string
              nullable: true
              enum: [sagittal, frontal, transverse, multi_planar]
            jointRegions:
              type: array
              items: { type: string }
            laterality:
              type: string
              nullable: true
              enum: [bilateral, unilateral, alternating, single_side]
            loadType:
              type: string
              nullable: true
              enum:
                [
                  bodyweight,
                  free_weight,
                  machine,
                  cable,
                  band,
                  cardio_machine,
                  assisted,
                  other
                ]
            skillType:
              type: string
              nullable: true
              enum:
                [strength, power, endurance, mobility, balance, coordination]
            flags:
              type: array
              items: { type: string }
            programming:
              type: object
              additionalProperties: true
              description: Free-form programming guidance, such as set and rep ranges.
            isPremium: { type: boolean }
            catalogVersion: { type: integer, minimum: 1 }
...
    SyncMetadata:
      type: object
      required: [catalogVersion, resources, endpoints, enums, pageSizeMax]
      properties:
        catalogVersion: { type: integer, example: 3 }
        latestExerciseUpdatedAt:
          type: string
          format: date-time
          nullable: true
        latestChangeAt:
          type: string
          format: date-time
          nullable: true
        resources:
          type: array
          items: { type: string }
          example: [metadata, exercises]
        endpoints:
          type: object
          additionalProperties: { type: string }
        enums:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
        pageSizeMax: { type: integer, example: 100 }
    Tombstone:
      type: object
      description: >
        A record that is not present in `exercises`. `deleted` means remove it
        locally; `deprecated` means flag it. `deprecated` tombstones are emitted
        only when `include_deprecated=false` — otherwise the record is returned
        in `exercises` with `status: deprecated` instead.
      required: [exerciseId, changeType, changedAt, catalogVersion]
      properties:
        exerciseId: { type: string, format: uuid }
        changeType:
          type: string
          enum: [deprecated, deleted]
        changedAt: { type: string, format: date-time }
        catalogVersion: { type: integer }
    SyncExercisesPage:
      type: object
      required: [exercises, tombstones, latestChangeAt]
      properties:
        exercises:
          type: array
          items: { $ref: '#/components/schemas/Exercise' }
        tombstones:
          type: array
          items: { $ref: '#/components/schemas/Tombstone' }
        latestChangeAt:
          type: string
          format: date-time
          nullable: true
          description: >
            Newest change in the catalog when this sync began. Identical on
            every page of one sync. Commit it after the final page and send it
            back as `updated_since`. Null when the catalog has never recorded a
            change.
          example: '2026-07-02T09:14:00.000Z'

    # ---------------------------------------------------------------------
    # Account models
    # ---------------------------------------------------------------------
    DeveloperAccount:
      type: object
      required: [id, email, tier, isAdmin, isActive, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string, nullable: true }
        tier:
          type: string
          enum: [free, basic, pro, enterprise]
        isAdmin: { type: boolean }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    ApiKeyMetadata:
      type: object
      required: [id, isActive, createdAt]
      properties:
        id: { type: string, format: uuid }
        label: { type: string, nullable: true, example: Mobile app }
        isActive: { type: boolean }
        lastUsedAt: { type: string, format: date-time, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    IssuedApiKey:
      description: An API key together with its plaintext value, returned exactly once.
      allOf:
        - $ref: '#/components/schemas/ApiKeyMetadata'
        - type: object
          required: [key]
          properties:
            key:
              type: string
              description: Store this now. It cannot be retrieved again.
              example: exdb_XEtT1oQ3n8yv2Kd6pR9wLz0aB4cF7hJmN1sT5uV8xYz
    AuthSession:
      type: object
      required: [user, apiKey]
      properties:
        user: { $ref: '#/components/schemas/DeveloperAccount' }
        apiKey: { $ref: '#/components/schemas/IssuedApiKey' }
    UsageDay:
      type: object
      required: [date, requestCount, limit, remaining]
      properties:
        date: { type: string, format: date, example: '2026-07-09' }
        requestCount: { type: integer, example: 143 }
        limit: { type: integer, example: 1000 }
        remaining: { type: integer, example: 857 }

    # ---------------------------------------------------------------------
    # Billing models
    # ---------------------------------------------------------------------
...
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, example: alice@example.com }
        password: { type: string, minLength: 8, example: a-long-passphrase }
        name: { type: string, example: Alice }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, example: alice@example.com }
        password: { type: string, example: a-long-passphrase }
    CreateApiKeyRequest:
      type: object
      properties:
        label: { type: string, maxLength: 100, example: Mobile app }
        expiresAt:
          {
            type: string,
            format: date-time,
            example: '2027-06-15T10:00:00.000Z'
          }
...
    AuthSessionResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/AuthSession' }
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [user]
              properties:
                user: { $ref: '#/components/schemas/DeveloperAccount' }
    LogoutResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [loggedOut]
              properties:
                loggedOut: { type: boolean, example: true }
    DeveloperAccountResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/DeveloperAccount' }
    ApiKeyResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/ApiKeyMetadata' }
    ApiKeyListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/ApiKeyMetadata' }
    IssuedApiKeyResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/IssuedApiKey' }
    UsageResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/UsageDay' }
...
    SyncMetadataResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required: [data]
          properties:
            data: { $ref: '#/components/schemas/SyncMetadata' }
    SyncExercisesResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          required: [data, pagination]
          properties:
            data: { $ref: '#/components/schemas/SyncExercisesPage' }
            pagination: { $ref: '#/components/schemas/CursorPagination' }
...
    ProblemDetail:
      type: object
      description: >-
        RFC 9457 Problem Details. Returned by every error response with content
        type `application/problem+json`.
      required:
        - type
        - title
        - status
        - code
      properties:
        type:
          type: string
          format: uri
          description: Stable URI identifying the error type. Does not resolve.
          example: https://exercisedb-api.dev/errors/validation-error
        title:
          type: string
          description: Short human-readable summary of the error type.
          example: Validation Error
        status:
          type: integer
          description: HTTP status code, matching the response status.
          example: 400
        detail:
          type: string
          description: >-
            Explanation specific to this occurrence. Replaced with a generic
            message for 5xx responses so internals are never exposed.
          example: limit must be less than or equal to 100
        instance:
          type: string
          description: URI of the request that produced the error.
          example: /exercises?limit=500
        code:
          type: string
          description: >-
            Machine-readable error code. RFC 9457 extension member retained for
            programmatic error handling.
          example: VALIDATION_ERROR
        requestId:
          type: string
          description: >-
            Correlation id for this request, also returned in the
            `X-Request-Id` response header. Quote it when reporting a problem:
            it is the only way to find the server log line behind a 5xx, whose
            `detail` is deliberately generic.
          example: 485fa7dd-03b1-44e1-ae56-6b88d13b652b